requirements.txt
python-dotenv
openai
chromium
git+https://github.com/openai/swarm.git
beautifulsoup4
tabulate